+++ crates/raven-database/src/import/mod.rs
@@ -1,8 +1,4 @@
-use std::{
-    fs::File,
-    io::{BufRead, BufReader, Error, Lines},
-    path::Path,
-};
+use std::io::Error as IoError;
 
 use crate::history::model::History;
 
@@ -10,6 +6,16 @@ pub mod zsh;
 
 #[derive(Debug)]
 pub struct ImportError;
+
+// Allow converting std::io::Error to ImportError
+impl From<IoError> for ImportError {
+    fn from(_: IoError) -> Self {
+        ImportError
+    }
+}
+
+#[derive(Debug)]
+pub struct LoadError;
 
 /// The importer handles parsing individual history items from an import source (such as a history
 /// file ), transforming them to `History` objects and passing them to the Loader to be persisted.
@@ -31,8 +37,6 @@ pub trait Importer: Sized {
     fn load(self, loader: &mut impl Loader) -> Result<(), ImportError>;
 }
 
-pub struct LoadError;
-
 /// The Loader handles persisting imported `History` objects into the raven
 /// database.
 pub trait Loader {
@@ -43,12 +47,4 @@ pub trait Loader {
     /// This function will return an error if the loader encounters an
     /// issue with persisting the `History` object.
     fn push(&mut self, hist: History) -> Result<(), LoadError>;
-}
-
-fn read_lines<P>(filename: P) -> Result<Lines<BufReader<File>>, Error>
-where
-    P: AsRef<Path>,
-{
-    let file = File::open(filename)?;
-    Ok(BufReader::new(file).lines())
 }